Improve watch app
